inserting food data to food tables implemented
