+++ app/src/main/java/edu/bu/fitnessfriend/fitnessfriend/add_food.java
@@ -10,8 +10,15 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 
+import java.text.DateFormat;
 import java.util.ArrayList;
+import java.util.Date;
+
+import edu.bu.fitnessfriend.fitnessfriend.database.foodDatabaseUtils;
+import edu.bu.fitnessfriend.fitnessfriend.database.myDatabaseHandler;
+import edu.bu.fitnessfriend.fitnessfriend.model.food;
 
 /**
  * Created by karan on 9/9/2017.
@@ -65,6 +72,25 @@ public class add_food extends AppCompatActivity{
     };
 
     protected void logFood(View view){
+
+        food foodObj = new food();
+
+        EditText foodNameField = (EditText) findViewById(R.id.food_name_input);
+        EditText qtyField = (EditText) findViewById(R.id.qty_input);
+        EditText foodCaloriesField = (EditText) findViewById(R.id.food_cal_input);
+
+        String foodName = foodNameField.getText().toString().trim();
+        double foodQuantity = Double.parseDouble(qtyField.getText().toString().trim());
+        double foodCalorie = Double.parseDouble(foodCaloriesField.getText().toString().trim());
+
+        foodObj.setFoodName(foodName);
+        foodObj.setQuantity(foodQuantity);
+        foodObj.setCalorie(foodCalorie);
+
+        myDatabaseHandler dbHandler = new myDatabaseHandler(getApplicationContext(),null,null,1);
+        foodDatabaseUtils foodDatabaseUtils = new foodDatabaseUtils(dbHandler);
+        foodDatabaseUtils.insertFood(foodObj);
+
         Snackbar logFoodSnackbar = Snackbar.make(findViewById(R.id.addFood), "Food Logged", 0);
         btnUtility.clearInputs(addFoodInputs);
         logFoodSnackbar.show();
